requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
pyyaml==6.0.1
aiofiles==23.2.1
watchdog==3.0.0
